requirements.txt
astroid==2.3.3
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
Click==7.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.2.13
django-allauth==0.39.1
django-appconf==1.0.3
django-bmemcached==0.2.4
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-leaflet==0.24.0
django-redis==4.10.0
django-redis-cache==2.0.0
django-select2==6.3.1
django-widget-tweaks==1.4.3
Faker==1.0.5
geojson==2.4.1
gunicorn==19.9.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.3
licenseheaders==0.6
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.17.4
oauthlib==3.0.1
psycopg2==2.7.6
psycopg2-binary==2.7.7
pycodestyle==2.5.0
pylibmc==1.6.0
pylint==2.4.4
python-binary-memcached==0.28.0
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.9
redis==3.2.1
requests==2.21.0
requests-oauthlib==1.2.0
reverse-geocoder==1.5.1
scipy==1.3.3
selenium==3.141.0
Shapely==1.6.4.post2
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
typed-ast==1.4.0
typing==3.6.6
uhashring==1.1
urllib3==1.24.2
virtualenv==16.4.3
Werkzeug==0.15.4
whitenoise==4.1.2
wrapt==1.11.2
django-phone-field==1.8.0
django-import-export==2.2.0
django-ckeditor==5.9.0